+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changes
 * [Developer] Removed several unused properties and classes to simplify data model.
 * [UI]: Analysis list page reworked to use Datatables.  Now does server side paging for performance increases.
 * [UI]: Replace large checkboxes on Project Samples page with standard platform dependant checkboxes.
+* [UI]: Allow deleting and cancelling an analysis.
 * [UI]: Updated pipeline selection interface by adding the pipeline colour to the background of the name and making the panels equal in height.
 * [UI]: Fixed issue where samples in the cart could not be sent to galaxy.